python dependency installation after venv activated
`virtualenv` doesn't activate the virtual environment after creation, so initial installation will fail. the `pip install  ...` command has been shifted after activation. this has the effect of always installing the requirements file, however the number of requirements are few and their sizes are small and they are skipped if already installed - the overhead would be minimal.